get value from href
